Backend/requirements.txt
Flask==3.0.0
Pillow==10.0.1
Requests==2.31.0
transformers==4.34.0